feat: 사용 중인 역 삭제 불가, 노선 삭제시 포함 구간도 전부 삭제
* 커버리지 테스트케이스 추가
* 필요한 컨트롤러 메서드에 트랜젝션 추가
* 리포맷 수행

+++ src/main/java/subway/service/StationService.java
@@ -30,7 +30,10 @@ public class StationService {
                 .collect(Collectors.toList());
     }
 
-    public boolean deleteLine(Long id) {
-        return stationDao.deleteById(id);
+    public boolean deleteStation(Long id) {
+        if (!stationDao.isUsingInSection(id)) {
+            return stationDao.deleteById(id);
+        }
+        return false;
     }
 }
